requirements.txt
# Core dependencies
openai>=1.43.0
python-dotenv>=1.1.0
PyYAML>=6.0.0

# For HTTP requests (required by openai)
httpx<0.28.0
httpcore>=1.0.9

# Optional: For future enhancements
# requests>=2.32.0  # For webhook integrations
# slack-sdk>=3.0.0  # For Slack bot integration
# jira>=3.0.0       # For Jira API integration
